[CNKAF-153] Rename the storage probe 'namespace' tag to "clusterid." (#1065)
For the first CNKAF-153 checkins, the storage probe was tagging stats with the "namespace" tag to  allows us to identify which cluster is having issues, as a given Kafka cluster has a specific K8s namespace. Unfortunately that tag isn't conformant with extant tags, so tracking it involves even more template variables on the dashboard. Switch the name of this tag from "namespace" to "clusterid" (which is commonly used in Kafka stats to identify the specific cluster).