KSTORAGE-339: Tiered storage system tests flaky due to tiering_completed check (#1050)
The tiered storage system tests perform a tiering_completed check which checks whether only one log segment exists for a topic partition on every broker. For the leader, this check was correct, but for the follower non-aligned segments may cause more than one segment to exist on for the same offset ranges. We now loosen this check and make sure that only one segment exists on one broker, and two or less exist on the other brokers. I hope we can do better in the future by checking whether the offset ranges for all follower segments overlap the leader's single segment, but this is difficult to achieve with metrics alone and I think we need better tooling for this kind of validation. I think we should fix the flakiness now and open a JIRA to improve the checks further.

We now also check whether tiering has started via a different mechanism. Rather than checking whether more than one log segment exists, we check whether the TierLogSize is non-zero.