CPKAFKA-4236: Check for empty observer list in addition to None (#1082)
* CPKAFKA-4236: Check for empty observer list in addition to None

If observers is not enabled (IBP < 2.4), then in
`alterPartitionReassignments` request we check that observers is not
provided. But the check only checks if Option is `None`, it doesn't
check if the Option is `Some(Seq.empty)`. This breaks the api as when
observers are not used with IBP < 2.4, the api returns with
INVALID_REPLICA_ASSIGNMENT error.

This change fixes it by checking for `None` as well as checking that the
value is not an empty list. Added a new unit test to make sure that we
check this case with IBP 2.3.

The change also updates code to return error as ApiError. Before this
change the method were returning Boolean to indicate failure/success.
This resulted in lost information on why failure occurred.