CPKAFKA-4247: resign as controller on KafkaController.shutdown (#1095)
We changed the shutdown order of the broker's components in 37aa84f. Previously the controller was shutdown just prior to shutdown of the zk client, which allowed the /controller node to be deleted and a new controller elected. Now shutdown of the log manager happens between these two events, which means that the time between controller resignation and removal of the /controller znode depends on the time to close all the existing logs (which involves flushing active segments and writing producer snapshots).

Instead of closing the zkClient immediately after controller shutdown, the controller now resigns on shutdown rather than relying on the zkClient shutdown to implicitly do so.