CPKAFKA-4178: replication system test should wait for hotset retention on local segments (#1099)
The replication system test is flaky because we wait for tiering to tier at least one segment, but we do not wait long enough for hotset retention to kick in before fetching. This results in no bytes being fetched from the tiered section of the log.

The solution here is to sleep for longer than log_retention_check_interval. I see no easy solution for this with respect to our current metrics, and this pattern is used elsewhere.