KSTORAGE-378: Add tier topic partition's information in TierMetadataState file's header (#1091)
Added extra FileTirePartitionState header's fields ‘localMaterializedOffset’ and ‘globalMaterializedOffset’.  The field ‘localMaterializedOffset’ will be used for determining the last tierTopicPartition’s event processed for the given metadata state. This offset will be used in rebuilding the state information and can be used in metadata integrity validation and recovery against corruption. The offset ‘globalMaterializedOffset’ is added as a place holder and subsequent PR, KSTORAGE-345 will use it for identifying the TierTopicPartition’s offset on bootstrap.
The unit test for these fields is added in TierPartitionStateTest, where they are tested for 1) default value, 2) update, 3) bootstrap, and 4) stale update.

Reviewers: @rite2nikhil, @dhruvilshah3