HOTFIX: Use default values when using v0 / v1 FileTierPartitionState (#1108)
This fixes a regression caused by 6b407d4 which prevents broker startup. The commit introduced v2 format of the tier partition state. We expected the new field added in the commit to always exist, causing an NPE when parsing the old header. This patch fixes the issue by correctly handling the absence of the old field and using default values in its place.