CPKAFKA-4272: revert changes to KafkaServer shutdown order to match AK (#1113)
Fixes CPKAFKA-4272 where a broker could rejoin the ISR after the broker's znode was removed. This resulted from a timing window created by the fact that the controller was shutdown before the replica fetchers, unlike Apache Kafka. ISR rejoin is able to occur due to a bug in epoch handling when using incremental fetch sessions (see KAFKA-9338).

This PR reverts many of the ordering differences compared to AK, and instead chooses to deal with the problematic dependency of the KafkaController on the TierTopicManager by allowing the TierTopicManager to be safely shutdown before the controller even though the controller depends on it.