KSTORAGE-388: Fix tier consumer registration for deleted logs (#1129)
We expect that only one of Log or TierDeletedPartitionsCoordinator is actively materializing the tier state of a particular partition. On log deletion, the Log layer unregisters itself to stop materializing the tier state. This allows TierDeletedPartitionsCoodintator to register and begin materialization if the partition was deleted.

Log deletion itself is an asynchronous process. We first rename the log directory to .deleted and schedule an async task to delete the segment files and the directory itself. If the broker restarts in between this duration, Log layer may re-register itself during log loading. This would then prevent TierDeletedPartitionsCoordinator from materializing the tier state. This PR fixes the issue by preventing registration for materialization from the Log layer if the log has already been deleted.