KSTORAGE-317: Delete stray partitions in log directory (#1034)
This PR adds a mechanism to delete stray partitions in the log directory. The first LeaderAndIsrRequest sent to a broker contains metadata for all replicas it hosts. We use this to determine if we have any stray partitions for which we are not a replica anymore, and if so, delete such partitions from broker's log directory. This patch also adds a new containsAllReplicas field to LeaderAndIsrRequest to make this operation safe, and to make the contract for the first LeaderAndIsrRequest more explicit.