CPKAFKA-4257. Check for empty topic placement constraint config (#1103)
* CPKAFKA-4257. Check for empty topic placement constraint config

If `confluent.placement.constraints` property is set to empty, we are
failing to create a topic with following error:

org.apache.kafka.common.errors.InvalidRequestException: Both replicationFactor and confluent.placement.constraints are set. Both cannot be used at the same time.

This change removes TopicPlacement.empty and checks for empty case and treats
it as null. The change also adds unit test for the error conditions so that in future we
can add more tests if needed, or just catch regressions.