MINOR: Add configuration to enable stray partition deletion (#1168)
This PR adds a broker configuration `confluent.enable.stray.partition.deletion` to control whether stray partition deletion is enabled. When disabled, we log a warning when we detect a stray partition but skip the actual deletion.

This PR also disables stray partition deletion by default, as this will help us validating the behavior before we enable it.