KAFKA-9491; Increment high watermark after full log truncation (#8037) (#1183)
* KAFKA-9491; Increment high watermark after full log truncation (#8037)

When a follower's fetch offset is behind the leader's log start offset, the
follower will do a full log truncation. When it does so, it must update both
its log start offset and high watermark. The previous code did the former,
but not the latter. Failure to update the high watermark in this case can lead
to out of range errors if the follower becomes leader before getting the latest
high watermark from the previous leader. The out of range errors occur when
we attempt to resolve the log position of the high watermark in DelayedFetch
in order to determine if a fetch is satisfied.

# ce-kafka
In ce-kafka the logStartOffset is maintained by MergedLog, so updateLogStartOffset has been renamed to maybeUpdateRecoveryPointAndHighwatermark and the code to set the Log.logStartOffset has been removed.

Reviewers: Rajini Sivaram <rajinisivaram@googlemail.com>, Chia-Ping Tsai <chia7712@gmail.com>, Ismael Juma <ismael@juma.me.uk>
Conflicts:
	core/src/main/scala/kafka/log/Log.scala

Co-authored-by: Jason Gustafson <jason@confluent.io>
Co-authored-by: Dhruvil Shah <dhruvil@confluent.io>