KSTORAGE-460: Disable preferred tier reads for replica fetch (#1226)
When replicas fetch from the leader and the offset has been tiered, the leader returns OFFSET_TIERED exception. The follower then attempts to find the start of the leader's local log using the ListOffsetRequest, materializes the tier partition state up to that point and continues replicating the leader's local log from that offset on.

With KSTORAGE-429, we enabled tier fetches for data that has been tiered but is still available in the local log. We accidentally enabled this path for fetches from replicas as well, which caused leaders to throw an OFFSET_TIERED exception. When the follower attempts to find the start of the leader's local log, it would realize that it may already have replicated that portion of the log, and would thus get stuck in a loop of receiving an OFFSET_TIERED exception / re-materializing tier partition state up to a point where it would again get an exception.

This PR disables preferred tier reads for replica fetches. These fetches will continue to be served as if preferred read was disabled.