KSTORAGE-337/KSTORAGE-348: tier deletion manager system test (#1115)
Adds a TierDeletionManager system test to check that all objects for a topic are deleted after topic deletion. Also adds "confluent.tier.s3.prefix": a static prefix to objects in S3. This allows us to perform a ListBucket on this prefix in the system test at the end of the test and check whether any objects exist that should not exist.