KSTORAGE-163: Initial implementation of TierFetcher memory bounding. (#1078)
* KSTORAGE-163: TierFetcher memory bounding.

The goal with this patch is to allow users to set a rough upper bound on the
amount of memory used by fetches from tiered storage. It's necessary to do this
for tiered storage due to how tiered fetches work.

Unlike fetches from local storage, tiered fetches are not chunked. The entire fetch
will be loaded into memory before being written to the socket. Because of this, it's
useful to have a rough upper bound on the amount of memory a fetch from tiered
storage can take.

The approach taken is to create 1 new type, `ReclaimableMemoryRecords`.
Along with adapting the existing `Send` impl to include `Send::release`, this
provides the ability to measure how much heap memory is in use after the initial
fetch from tiered storage up until the `Send` is written to the socket.

When a fetch from tiered storage is requested, we
first check the see if the memory tracker pool size is non zero. If
it is non zero, we allow fetching the first batch and returning
`ReclaimableMemoryRecords`. If the pool size is less than or equal to
zero, a `TierFetcher` thread is blocked. This is intended to be a compromise
between keeping a tight bound on memory consumption, and avoiding wasted
work.

`ReclaimableMemoryRecords` are then completed as normal after purgatory,
except additional logic is put in place to explicitly reclaim memory
if a `ReclaimableMemoryRecords` instance is dropped. Additional notes
have been added to function preludes indicating this requirement.

Ultimately, during construction of the response,
`ReclaimableMemoryRecords::toSend` will return a `Send` implementation
which offers a passthrough `Send::release` method for releasing the memory
leased by the `ReclaimableMemoryRecords`.