Define a DataBalancer interface and object in the Kafka Controller. (#1218)
The trait and object are intended as a facade for the eventual inclusion of CruiseControl,
but nicely hiding and abstracting away all of CruiseControl's cruft. It
also provides a long-lived object for Kafka interactions to take place
(e.g. dynamic variable configuration, API endpoints).

This implements the DataBalancer trait/interface to identify needed functionality and a concrete KafkaDataBalancer object, currently minimal, to actually implement that functionality.