CNKAF-323: Add metric with per-tenant partition count (#1241)
This commit adds a per-tenant partition count metric. The metric is exposed by every broker and exposed the total number of partition per tenant.

As the CreateTopicPolicy already maintains the number of partition per tenant, we have stretched it a bit further to expose it via a metric as well.