KCORE-111: Default replica placement constraints for external topics (#1277)
Implements broker config for topic/log placement constraint. This implementation is different from the normal broker specified log configurations. For placement constraints, Kafka will persist the value specified in the broker at topic creation time. Deleting the placement constraint for a topic doesn't cause the value of the placement constraint to revert back to the broker value. It instead reverts to no placement constraint being specified for the topic.