[CNKAF-283] Put heavily trimmed CruiseControl into ce-kafka/ce-sbk (#1331)
* Copy-pasta a lot of Cruise Control Files directly into the ce-kafka repository.

Do the copy-paste instead of a subtree import so that it fits into the ce-kafka core/ tree
as an import.

Note that the async and (most of) the servlet and metricsreporter subtrees are *not* imported,
as those are not relevant for in-broker cruise-control.

* Add CruiseControl to gradle build.

* Add Apache Commons IO library.

* Remove Executor reliance on the UserTaskManager.

This removes the ability to report and distinguish user-initiated tasks.

* Remove unnecessary import of CruiseControlMetricsReporter.

* Remove unnecessary import.

* Remove ConfluentMetricsReporterSampler.java. Doesn't belong in core build.

Will migrate in elsewhere later.

* Remove YammerMetricsProcessor.java.

* Remove CruiseControlStateParameters.

* Remove Executor dependence on UserTaskManager: tests.

* Remove call to UserRequestException. Replace with a generic Exception(!). Needs review.

Need to decide what the right solution is here.

* Remove CruiseControlParametersConfig and references to it.

* CruiseControConfig: Remove reference to MetricReporter config.

* Remove CruiseControlRequestConfig.

* Remove KafkaCruiseControlMain. Program startup isn't needed in-broker.

* Update Kafka SampleStore to use modern adminZkClient interfaces.

* Switch KafkaCruiseControl from throwing a UserTaskException to just an Exception. Needs review.

XXX: Pick which exception to really throw.

* Remove reliance on a non-present Parameter value (Rack-Aware-Checking). Change thrown Exception.

* Remove dependencies on AbstractCruiseControlResponse for BrokerStats, CruiseControlState.

This involves amputating some (but not all) state.

These classes were passed to the outside world and therefore had a bunch of stuff to be translated into
text or JSON. Remove those dependencies so we can use them in a servlet-less Cruise Control.

* Remove servlet-based dependencies of the KafkaCruiseControlConfig.

* Remove dependencies on MetricReporterConfig from the Sampler.

Specifically the default topic and the producer config (used to determine max tolerable latency).

* Remove ConfluentMetricsReporterSamplerTest.

* AdminZK API update for LoadMonitor.test.

* Add OptimizationResult class.

* Test utilities for CC tests.

* TECH DEBT: Lots of suppressions and nastiness to allow Cruise Control to build.

Cruise Control has lots of style divergences and other issues which cause the ce-kafka build to fail.
Add Checkstyle suppressions and remove '-Werror' while this code imports.

* Fix resource leakage of ReplicationThrottleHelperTest.

* Fix resource leakage, KafkaSampleStoreTest.

* Fix resource leakage, LoadMonitorTaskRunner test.

* Werror fixes: EasyMock casting issues.

* Werror fixes: Replace ExpectedException rules with assertThrows (JUnit4->5).

* PAY TECH DEBT: Werror fixes: Re-enable Werror for core builds.

* Build hackery: Exclude Executor tests from running as they break other tests.

* Werror fixes: Remove ExpectedException from Executor Test.

* Fix resource leakage, ExecutorTest.

* PAY TECH DEBT: Allow all tests to run.

* "Fix" notification-after-user-task test.

Modify the notification-after-user-task test to work with the fact that
user notification types are not generated now.
 (Really should be an expected fail but I couldn't find that?)

* Review feedback: For the generic Exception in KafkaCruiseControl, note bug and TODO.

* Review feedback: Clean up the OptimizationResult class.

This class was imported and amputated from CruiseControl, but we can at least remove
random commenting-out of code.

* ZkClient API changes due to Kafka update.

Stop using internal APIs!

* CruiseControl test fix: DummyLicenseValidator: needed to make fake brokers.

* Gradle changes for CE-SBK

* Add ce-sbk as a gradle subproject.

* SBK checkstyle: Reenable"Unused imports" checking.

* SBK checkstyle: Remove "No Whitespace" suppression.

* SBK checkstyle: Remove "Unnecessary Parentheses" suppression.

* SBK checkstyle: Remove "AvoidStarImports" suppression.

* Turn off Regexp checkstyle issue.

THIS INVOLVES CODE CHANGES -- the code was using non-locale-aware string ops, and one of these
was parsing an exception string. Replace this with a better method, checking for the explicit
CorruptRecordException.

* SBK checkstyle: Document reasoning behind a number of checkstyle suppressions.

* Remove the DummyLicenseValidator copy.

Use Gradle magic to make the object available. Whew.

* build fixes/updates

* Remove unneeded META-INF -- now inherited from core.

* Flaky tests fix: Don't use real time when you're using a mock timer!

Some AnomalyDetectorTests were passing in system time when the test was using a mock timer,
and on slow systems these values can diverge. When using timestamps in a test, always pass
in the time from the mock timer.

* Build fixes: Properly name the CE-SBK jar file, make releaseTar depend on ce-sbk libs as well.

* @Ignore two flaky tests with race conditions (CNKAF-450)

Co-authored-by: Vikas Singh <vikas@confluent.io>