KSTORAGE-576  Ignore incrementing endOffset marker for fenced mapping in TierMetadataValidator
In TierMetadataValidator, while doing deep validation against offset ranges, we see a case where under active offset range, the validator will increment the end offset marker for all the continuous segments, even if the state of that segment is Fenced. As a result of this offset increment, on receiving the next valid segment, there will be an end offset marker mismatch which raises the false alarm of inconsistencies. The fix is to ignore the incrementing endOffset marker.
Added unit test for the same.

reviewer @soumym