[CNKAF-477] Restore ConfluentMetricsReporterSampler (#1391)
The checkin of CruiseControl into the ce-sbk directory did not include either the CruiseControlMetricsReporter nor the ConfluentMetricsReporterSampler. The former we don't use because we use the ConfluentMetricsReporter to get metrics; the latter was simply omitted.

This change imports the ConfluentMetricsReporterSampler and associated unit test
into ce-sbk. There are no changes to the code from that in the Confluent CruiseControl repository.