CPKAFKA-4670 Fix for system tests running 0.8.2.2 version producers (#1396)
Fix for system tests running 0.8.2.2 version producers

VerifiableProducer.java is a wrapper class around KafkaProducer and is
used for running system tests. It was recently modified to provide timestamp
for the records produced. For clients version 0.8.2.2, RecordMetadata class
did not provide a timestamp() method and hence some system tests started to
fail for these clients.
This change will call the older way of reporting timestamp, if
RecordMetadata does not have the timestamp() method.