KSTORAGE-568: bound hotset retention by LSO (#1384)
In KSTORAGE-568 the following pattern occurred. A hung transaction caused the LSO to not advance. This may or may not have been triggered by KAFKA-9144.

The leader did not appear to have a hung transaction, and archived the segment. Then the follower materialized the object metadata and deleted it from its hotset. Later the follower fetched from the leader which caused Log.maybeIncrementLogStartOffset to be called. As the FSO was inside the tiered section of the log, the local segment that it corresponded to could not be found.

In this change we bound hotset retention to the LSO so that we do not delete local segments until the LSO advances.