KSTORAGE-399 : Crash broker if FileTierPartitionState encounters corrupted checkpointed state file during initialization. (#1394)
Crash broker if the FileTierPartitionState encounters corrupted flushed state file during initialization. This will prevent any potential data loss because in the absence of tier state information the logStartOffset may be incremented leading to loss of data. There are improvements being worked on and till then it is necessary to avoid any data loss. The broker can be resumed after manual recovery.

Added unit test for the same and also simulated and tested the behavior in the end-to-end test.

Reviewers: @dhruvilshah3 @kowshik