[CNKAF-216] Derive SBK capacity values from config parameters directly. (#1416)
CruiseControl stores broker capacity information, allowed to be specialized per-broker, in
a file that is expected to be loaded from local disk. This capacity configuration is expected
to provide CPU, disk, and network in/out capacities.

Self-Balancing Kafka is expecting to obtain this information from config parameters, generalized for all brokers in the cluster, as specified in https://confluentinc.atlassian.net/wiki/spaces/CNKAF/pages/1064345722/SBK+UX+one-pager?focusedCommentId=1137548431#comment-1137548431.

This change implements a BrokerCapacityFromConfigResolver class which gets capacity information from configuration properties, expected to be strings. If these values are invalid, a ConfigException is thrown.

 Because only default  (not per-broker) capacity  is tracked, the BrokerCapacityInfo is made somewhat immutable and the accessor methods now return unmodifiable objects.

As CPU capacity is not considered to be an issue for SBK, the CPU capacity defaults to 100%.

Note that this change requires disk capacity to be specified but that is to be removed by https://confluentinc.atlassian.net/browse/CNKAF-372, which will make capacity detection automatic.