KSTORAGE-453: Persist leader epoch in tier checkpoint file (#1220)
This PR adds leader epoch corresponding to the last consumed position to the tier checkpoint file and FileTierPartitionState. The added protections introduced in this PR are applicable only when using v2 message format or higher.

The epoch is used in couple of ways:

To seek the consumer to the last committed leaderEpoch in addition to the offset so the consumer can detect whether the tier topic was truncated.
To detect cases where the leader epoch is inconsistent with the last epoch materialized in FileTierPartitionState or the committed epoch checkpoint file.
When an illegal epoch is detected by the consumer or when updating the tier offsets checkpoint file, we will stop the TierTopicConsumer / TierTopicManager thread, depending on when the exception was raised. The effect of this is to stop any further archiving on the broker.

When an illegal epoch is detect when appending to the tier partition state file, we end up fencing the partition and move it to the ERROR state, which will stop further archiving for the specific partition.

Note that while there is a change in the persistent format which we would ideally put behind an IBP bump, we do not currently have a mechanism to version proprietary formats. This is not an issue for now as we are tolerant to unknown versions, and handle this case by resetting the offsets to beginning. i.e. we have the ability to rollback to a version without this patch.