CNKAF-520, CNKAF-525: Wait for KafkaSampleStore check to complete at startup (#1454)
KafkaSampleStore checks that the two topics that it uses to store
historical data are present at startup. If it finds any issue it aborts
the Cruise Control startup. It can run into error for many reasons but
the one common is that topic replication factor isn't met. This is
because during bootstrap first broker will become controller, it will
try to start CC and then SampleStore will complain that topic
replication factor is 3 whereas there are only 1 alive brokers and then
abort the CC startup. This makes it difficult to start brokers and have
CC running.

This change refactors the code to add the check as part of CC startup
check. Tested by running broker in different setups to confirm that it
works.