CNKAF-548: Skip checking quotas when recording replication traffic (#1480)
It was found that checking for quotas in throttled replicas accounted for 5% of a CPU profile when all replicas' throttle was enabled. The majority of the time is spent in constructing the QuotaViolationException (stacktrace + String construction). Given that the exception is purposefully ignored, it makes sense to skip its generation when recording quotas for replication traffic