CNKAF-50: Do not set topic-level throttled.replicas if broker static throttling is enabled (#80)
As of ce-kafka:v0.810.0-6.0.0-ce-SNAPSHOT, we have static replication throttling in Kafka. This feature makes it so that Kafka can enable replication throttling for all replicas on a broker.

Dynamic topic-level throttling overrides these in the sense that it may unthrottle some replicas of a topic if only a subset of all are throttled in the topic-level. This patch changes Cruise Control to not set these topic-level throttles if each broker already has the static replication throttles enabled