CNKAF-541: Reduce the number of tests run in RandomClusterTest (#1474)
CNKAF-541: Reduce the number of tests run in RandomClusterTest

Subclasses of RandomClusterTess were running 48 tests each. The tests
were groups in 4 sets of 6x2 tests. This change reduces the number to 4
sets of 4x2 tests. In all total number of tests run is now 32. The
reduction in time it takes for the tests to run is about 50% (6m 39s to
3m 9s) as the indexes were used to generate # of partitions/replica etc,
thus getting rid of higher numbers resulted in getting rid of longer
running tests.

The change also categorizes the test as IntegrationTest so it will not
get run if unitTest goal is used.