MINOR: Add check for rack method to Partition class (#1473)
The original partitionRacks method returned list of racks in the
partition. The return value was only used to check if the rack was
present for partition. The creation of set by iterating over partition
replica was consuming about 15% time when running `RandomClusterTest`
subclass tests.

This change removes the method and instead adds a method to just perform
the check if a rack belongs to one of the replica in the partition. The
check is lazy and doesn't go over all replicas if a match is found. The
change saved about 10% of runtime for test.