MINOR: FileTierPartitionState refactor statefulness into inner class to support recovery (#1465)
This is a refactor of FileTierPartitionState, intended to encapsulate almost all of the state that will need to be swapped on recovery into its inner State class. This will allow our recovery process to load a new FileTierPartitionState file and perform a swap across the states in an atomic way, and will does not require cleaning up the object.

Example of such state includes uploadInProgress, which if improperly reset would cause a new divergence. I have used the opportunity to also encapsulate more methods into this class which were somewhat awkwardly called from outside of it with state as an argument.

By separating this out cleanly we will eventually be able to download a new state from object storage, scan it and create a new FileTierPartitionState, and only then will we have to take a lock out that could prevent metadata lookups.