KSTORAGE-626: Remove metrics reporter configuration when configuring internal tier topic clients (#1503)
A side effect of 52e1ed1 is that it copies over unneeded broker configurations when configuring tier topic clients. We found a problem with the client using the broker's configured metrics reporter, which could lead duplicate and incorrect metrics in the metrics topic. This PR ensures that metrics reporter configurations are not copied over into tier topic clients.