[CNKAF-582] Handle topic creation during checkStartupComponents (#1526)
During KafkaSampleStore startup, if a SampleStore topic was created during the checkStartupComponents phase, topic creation might not have fully completed and resulted in an error.

Modify the KafkaSampleStore::ensureTopicCreated()  (renamed to checkTopicExistsMaybeCreate()) code path to return true if the desired SampleStore topic exists and is accessible
(even if not fully configured) and false if the topic had to be created. By doing so, there will be another loop to validate topic existence in SampleStore startup, which allows actual startup to proceed without errors.

Add a unit test for this code path.