CNKAF-585; Multi tenant interceptor rejects create topic requests with NO_REPLICATION_FACTOR or NO_NUM_PARTITIONS when quotas are enabled (#1552)
We have found out that the multi tenant interceptor rejected create topic requests which use the default replication factor and the default number of partitions, both represented by -1 in the request, when a partition assignor is used. This is the case when quotas are enabled.

Concretely it means that creating a topic like this kafka-topics.sh --bootstrap-server ... --create --topic cc-test fails with InvalidRequestException: Invalid partition count -1.

This PR updates the MultiTenantRequestContext to use the default replication factor and the default number of partitions to generate the assignments when a partition assignor is used. The PR has been tested in cpd.