[CNKAF-594] Support broker shutdown for SBK (#1569)
* Rename the DataBalanceManager::{startUp, shutdown} methods to {onElection, onResignation}.

This more closely mimics what the related Controller code is called, and what the actual
calling context is expected to be.

* Make the DataBalanceEngine handle shutdown vs deactivation.

Deactivation is when the DatabalanceEngine should stop executing but may be reactivated
at some point in the future (i.e. it goes quiescent).
Shutdown is when we know that the object is done executing.

Enable/disable activates and deactivates the DataBalanceEngine, but Manager resignation is
an actual shutdown (as the DataBalanceEngine is released). Call DataBalanceEngine::shutdown
during KafkaDataBalanceManager resignation.

* Make the active and inactive DataBalanceEngines cached.

Have the DataBalanceEngineFactory cache instances of the Active and Inactive DataBalanceEngines.
Implement a shutdown() routine for broker shutdown.