MINOR: triage upgrade system test for 5.4 tiered storage upgrade (#1575)
There is a problem with how we collect metrics in the upgrade system test. Many of the metrics we need in check_cluster_state are not available in 5.4 and I have had no luck adding them after the roll for now.

This PR triages this issue and also improves the final tier status check to check tiering_completed rather than tiering_started if the hotset is reduced.

I have been hitting this problem while testing #1471, and bisected until it showed de9cb7d to be at fault. I think we should merge this fix rather than block on getting the full metrics scraping working. I have opened https://confluentinc.atlassian.net/browse/KSTORAGE-653 so we remember to re-add the check_cluster_state check.