SOAK: Revise soak workloads (#1597)
We want to enable the new unbalanced workload in our soak cluster. Because it's overloaded we need to lower some of the existing workloads’ throughput in order to accommodate for this new one.

This patch:

- Removes enterprise-topics’ 1600 partitions test
- Increases one-large-topic’s partitions to 1600. Increase it’s produce to 70MB/s
- Removes the high-load-topic workload.

We're basically consolidating the high-load-topic and enterprise topics' many partitions test into one-large topic and introducing the new unbalanced workload in our soak cluster.

With all of this, we should have a scheduled 146MB/s of throughput on the cluster - the same as what we currently see in practice - 152MB/s.