[CNKAF-606] Stop proposal execution on Cruise Control shutdown/SBK deactivation (#1620)
Make sure the Executor comes to a stop on SBK deactivation. This applies for both controller
resignation as well as SBK disable. If self-healing is on, it's OK since on restart the
system will notice imbalances and compute a new rebalance plan.

This sets the stopExecution flag way down in the Executor for ongoing executions, which on the next execution check will result in reassignments being canceled, and no further reassignments will be submitted.

Prior to this, CruiseControl shutdown would wait(!!) for ongoing executions to finish(!!).