MINOR: Introduce new ConfluentAdmin#create public API and improve SBK AdminClient testability (#1636)
This patch introduces a way to create a ConfluentAdmin client via create() methods, similar to how we have Admin#create()

It also switches ExecutorAdminUtils from a static helper class to an instance. This will greatly improve testability and maintainability of tests by reducing the massive amounts of boilerplate code we would need to do otherwise to mock the nitty-gritty AdminClient logic. It's a necessary step in making classes like the Executor unit-testable

Also updates various Cruise Control code to use a ConfluentAdmin instance