CNKAF-622: Properly handle dynamic config changes. (#1661)
KafkaDataBalancer::updateConfig would take only one KafkaConfig (the "new" config) under the assumption that the configuration stored in the KafkaDataBalancer properly reflected the old config. It
turns out that the "old" config is dynamically updated and the "old" config passed to the
dynamic broker reconfiguration routine is actually the copy (i.e. the old copy is more necessary
than the new one).

Instead treat this as the reconfigure routine used by all other dynamic configs and pass in
the old, new config, treating them as totally new.